Excepts the OPTIONS http method from the @authenticated decorator.
Before a browser will do a cross-origin POST or PUT, it does an OPTIONS check first.
It won't pass its cookies to the OPTIONS call, so Dorthy can't tell that the client
is in fact authenticated. This stops the browser from issuing any POST requests
to endpoints protected by @authenticated in the case of a cross-origin setup.